restructure; add support for enums
